requirements.txt
bilibili-api-python>=17.0
httpx>=0.27
rich>=13.0
questionary>=2.0
av>=12.0
mutagen>=1.47
Pillow>=10.0
